+++ crates/matrix-sdk-base/src/store/integration_tests.rs
@@ -26,7 +26,7 @@ use ruma::{
     },
     mxc_uri, room_id,
     serde::Raw,
-    uint, user_id, EventId, OwnedEventId, RoomId, UserId,
+    uint, user_id, EventId, OwnedEventId, OwnedUserId, RoomId, UserId,
 };
 use serde_json::{json, value::Value as JsonValue};
 
@@ -73,6 +73,10 @@ pub trait StateStoreIntegrationTests {
     async fn test_stripped_non_stripped(&self) -> Result<()>;
     /// Test room removal.
     async fn test_room_removal(&self) -> Result<()>;
+    /// Test presence saving.
+    async fn test_presence_saving(&self);
+    /// Test display names saving.
+    async fn test_display_names_saving(&self);
 }
 
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
@@ -182,7 +186,7 @@ impl StateStoreIntegrationTests for DynStateStore {
             )]),
         );
 
-        changes.add_stripped_room(stripped_room);
+        changes.add_room(stripped_room);
 
         let stripped_member_json: &JsonValue = &test_json::MEMBER_STRIPPED;
         let stripped_member_event = Raw::new(&stripped_member_json.clone()).unwrap().cast();
@@ -306,12 +310,10 @@ impl StateStoreIntegrationTests for DynStateStore {
 
         assert!(self.get_kv_data(StateStoreDataKey::SyncToken).await?.is_some());
         assert!(self.get_presence_event(user_id).await?.is_some());
-        assert_eq!(self.get_room_infos().await?.len(), 1, "Expected to find 1 room info");
-        assert_eq!(
-            self.get_stripped_room_infos().await?.len(),
-            1,
-            "Expected to find 1 stripped room info"
-        );
+        assert_eq!(self.get_room_infos().await?.len(), 2, "Expected to find 2 room infos");
+        #[allow(deprecated)]
+        let stripped_rooms = self.get_stripped_room_infos().await?;
+        assert_eq!(stripped_rooms.len(), 1, "Expected to find 1 stripped room info");
         assert!(self
             .get_account_data_event(GlobalAccountDataEventType::PushRules)
             .await?
@@ -376,22 +378,100 @@ impl StateStoreIntegrationTests for DynStateStore {
     async fn test_member_saving(&self) {
         let room_id = room_id!("!test_member_saving:localhost");
         let user_id = user_id();
+        let second_user_id = user_id!("@second:localhost");
+        let third_user_id = user_id!("@third:localhost");
+        let unknown_user_id = user_id!("@unknown:localhost");
 
+        // No event in store.
+        let mut user_ids = vec![user_id.to_owned()];
         assert!(self.get_member_event(room_id, user_id).await.unwrap().is_none());
+        let member_events = self
+            .get_state_events_for_keys_static::<RoomMemberEventContent, _, _>(room_id, &user_ids)
+            .await;
+        assert!(member_events.unwrap().is_empty());
+        assert!(self.get_profile(room_id, user_id).await.unwrap().is_none());
+        let profiles = self.get_profiles(room_id, &user_ids).await;
+        assert!(profiles.unwrap().is_empty());
+
+        // One event in store.
         let mut changes = StateChanges::default();
+        let raw_member_event = membership_event();
+        let profile = raw_member_event.deserialize().unwrap().into();
         changes
             .state
             .entry(room_id.to_owned())
             .or_default()
             .entry(StateEventType::RoomMember)
             .or_default()
-            .insert(user_id.into(), membership_event().cast());
-
+            .insert(user_id.into(), raw_member_event.cast());
+        changes.profiles.entry(room_id.to_owned()).or_default().insert(user_id.to_owned(), profile);
         self.save_changes(&changes).await.unwrap();
-        assert!(self.get_member_event(room_id, user_id).await.unwrap().is_some());
 
+        assert!(self.get_member_event(room_id, user_id).await.unwrap().is_some());
+        let member_events = self
+            .get_state_events_for_keys_static::<RoomMemberEventContent, _, _>(room_id, &user_ids)
+            .await;
+        assert_eq!(member_events.unwrap().len(), 1);
         let members = self.get_user_ids(room_id, RoomMemberships::empty()).await.unwrap();
-        assert!(!members.is_empty(), "We expected to find members for the room")
+        assert_eq!(members.len(), 1, "We expected to find members for the room");
+        assert!(self.get_profile(room_id, user_id).await.unwrap().is_some());
+        let profiles = self.get_profiles(room_id, &user_ids).await;
+        assert_eq!(profiles.unwrap().len(), 1);
+
+        // Several events in store.
+        let mut changes = StateChanges::default();
+        let changes_members = changes
+            .state
+            .entry(room_id.to_owned())
+            .or_default()
+            .entry(StateEventType::RoomMember)
+            .or_default();
+        let changes_profiles = changes.profiles.entry(room_id.to_owned()).or_default();
+        let raw_second_member_event =
+            custom_membership_event(second_user_id, event_id!("$second_member_event"));
+        let second_profile = raw_second_member_event.deserialize().unwrap().into();
+        changes_members.insert(second_user_id.into(), raw_second_member_event.cast());
+        changes_profiles.insert(second_user_id.to_owned(), second_profile);
+        let raw_third_member_event =
+            custom_membership_event(third_user_id, event_id!("$third_member_event"));
+        let third_profile = raw_third_member_event.deserialize().unwrap().into();
+        changes_members.insert(third_user_id.into(), raw_third_member_event.cast());
+        changes_profiles.insert(third_user_id.to_owned(), third_profile);
+        self.save_changes(&changes).await.unwrap();
+
+        user_ids.extend([second_user_id.to_owned(), third_user_id.to_owned()]);
+        assert!(self.get_member_event(room_id, second_user_id).await.unwrap().is_some());
+        assert!(self.get_member_event(room_id, third_user_id).await.unwrap().is_some());
+        let member_events = self
+            .get_state_events_for_keys_static::<RoomMemberEventContent, _, _>(room_id, &user_ids)
+            .await;
+        assert_eq!(member_events.unwrap().len(), 3);
+        let members = self.get_user_ids(room_id, RoomMemberships::empty()).await.unwrap();
+        assert_eq!(members.len(), 3, "We expected to find members for the room");
+        assert!(self.get_profile(room_id, second_user_id).await.unwrap().is_some());
+        assert!(self.get_profile(room_id, third_user_id).await.unwrap().is_some());
+        let profiles = self.get_profiles(room_id, &user_ids).await;
+        assert_eq!(profiles.unwrap().len(), 3);
+
+        // Several events in store with one unknown.
+        user_ids.push(unknown_user_id.to_owned());
+        let member_events = self
+            .get_state_events_for_keys_static::<RoomMemberEventContent, _, _>(room_id, &user_ids)
+            .await;
+        assert_eq!(member_events.unwrap().len(), 3);
+        let profiles = self.get_profiles(room_id, &user_ids).await;
+        assert_eq!(profiles.unwrap().len(), 3);
+
+        // Empty user IDs list.
+        let member_events = self
+            .get_state_events_for_keys_static::<RoomMemberEventContent, OwnedUserId, _>(
+                room_id,
+                &[],
+            )
+            .await;
+        assert!(member_events.unwrap().is_empty());
+        let profiles = self.get_profiles(room_id, &[]).await;
+        assert!(profiles.unwrap().is_empty());
     }
 
     async fn test_filter_saving(&self) {
@@ -448,8 +528,21 @@ impl StateStoreIntegrationTests for DynStateStore {
     async fn test_stripped_member_saving(&self) {
         let room_id = room_id!("!test_stripped_member_saving:localhost");
         let user_id = user_id();
+        let second_user_id = user_id!("@second:localhost");
+        let third_user_id = user_id!("@third:localhost");
+        let unknown_user_id = user_id!("@unknown:localhost");
 
+        // No event in store.
         assert!(self.get_member_event(room_id, user_id).await.unwrap().is_none());
+        let member_events = self
+            .get_state_events_for_keys_static::<RoomMemberEventContent, _, _>(
+                room_id,
+                &[user_id.to_owned()],
+            )
+            .await;
+        assert!(member_events.unwrap().is_empty());
+
+        // One event in store.
         let mut changes = StateChanges::default();
         changes
             .stripped_state
@@ -458,12 +551,67 @@ impl StateStoreIntegrationTests for DynStateStore {
             .entry(StateEventType::RoomMember)
             .or_default()
             .insert(user_id.into(), stripped_membership_event().cast());
-
         self.save_changes(&changes).await.unwrap();
-        assert!(self.get_member_event(room_id, user_id).await.unwrap().is_some());
 
+        assert!(self.get_member_event(room_id, user_id).await.unwrap().is_some());
+        let member_events = self
+            .get_state_events_for_keys_static::<RoomMemberEventContent, _, _>(
+                room_id,
+                &[user_id.to_owned()],
+            )
+            .await;
+        assert_eq!(member_events.unwrap().len(), 1);
         let members = self.get_user_ids(room_id, RoomMemberships::empty()).await.unwrap();
-        assert!(!members.is_empty(), "We expected to find members for the room")
+        assert_eq!(members.len(), 1, "We expected to find members for the room");
+
+        // Several events in store.
+        let mut changes = StateChanges::default();
+        let changes_members = changes
+            .stripped_state
+            .entry(room_id.to_owned())
+            .or_default()
+            .entry(StateEventType::RoomMember)
+            .or_default();
+        changes_members
+            .insert(second_user_id.into(), custom_stripped_membership_event(second_user_id).cast());
+        changes_members
+            .insert(third_user_id.into(), custom_stripped_membership_event(third_user_id).cast());
+        self.save_changes(&changes).await.unwrap();
+
+        assert!(self.get_member_event(room_id, second_user_id).await.unwrap().is_some());
+        assert!(self.get_member_event(room_id, third_user_id).await.unwrap().is_some());
+        let member_events = self
+            .get_state_events_for_keys_static::<RoomMemberEventContent, _, _>(
+                room_id,
+                &[user_id.to_owned(), second_user_id.to_owned(), third_user_id.to_owned()],
+            )
+            .await;
+        assert_eq!(member_events.unwrap().len(), 3);
+        let members = self.get_user_ids(room_id, RoomMemberships::empty()).await.unwrap();
+        assert_eq!(members.len(), 3, "We expected to find members for the room");
+
+        // Several events in store with one unknown.
+        let member_events = self
+            .get_state_events_for_keys_static::<RoomMemberEventContent, _, _>(
+                room_id,
+                &[
+                    user_id.to_owned(),
+                    second_user_id.to_owned(),
+                    third_user_id.to_owned(),
+                    unknown_user_id.to_owned(),
+                ],
+            )
+            .await;
+        assert_eq!(member_events.unwrap().len(), 3);
+
+        // Empty user IDs list.
+        let member_events = self
+            .get_state_events_for_keys_static::<RoomMemberEventContent, OwnedUserId, _>(
+                room_id,
+                &[],
+            )
+            .await;
+        assert!(member_events.unwrap().is_empty());
     }
 
     async fn test_power_level_saving(&self) {
@@ -740,7 +888,9 @@ impl StateStoreIntegrationTests for DynStateStore {
     async fn test_persist_invited_room(&self) -> Result<()> {
         self.populate().await?;
 
-        assert_eq!(self.get_stripped_room_infos().await?.len(), 1);
+        #[allow(deprecated)]
+        let stripped_rooms = self.get_stripped_room_infos().await?;
+        assert_eq!(stripped_rooms.len(), 1);
 
         Ok(())
     }
@@ -751,7 +901,9 @@ impl StateStoreIntegrationTests for DynStateStore {
 
         assert!(self.get_member_event(room_id, user_id).await.unwrap().is_none());
         assert_eq!(self.get_room_infos().await.unwrap().len(), 0);
-        assert_eq!(self.get_stripped_room_infos().await.unwrap().len(), 0);
+        #[allow(deprecated)]
+        let stripped_rooms = self.get_stripped_room_infos().await?;
+        assert_eq!(stripped_rooms.len(), 0);
 
         let mut changes = StateChanges::default();
         changes
@@ -768,21 +920,25 @@ impl StateStoreIntegrationTests for DynStateStore {
             self.get_member_event(room_id, user_id).await.unwrap().unwrap().deserialize().unwrap();
         assert!(matches!(member_event, MemberEvent::Sync(_)));
         assert_eq!(self.get_room_infos().await.unwrap().len(), 1);
-        assert_eq!(self.get_stripped_room_infos().await.unwrap().len(), 0);
+        #[allow(deprecated)]
+        let stripped_rooms = self.get_stripped_room_infos().await?;
+        assert_eq!(stripped_rooms.len(), 0);
 
         let members = self.get_user_ids(room_id, RoomMemberships::empty()).await.unwrap();
         assert_eq!(members, vec![user_id.to_owned()]);
 
         let mut changes = StateChanges::default();
         changes.add_stripped_member(room_id, user_id, custom_stripped_membership_event(user_id));
-        changes.add_stripped_room(RoomInfo::new(room_id, RoomState::Invited));
+        changes.add_room(RoomInfo::new(room_id, RoomState::Invited));
         self.save_changes(&changes).await.unwrap();
 
         let member_event =
             self.get_member_event(room_id, user_id).await.unwrap().unwrap().deserialize().unwrap();
         assert!(matches!(member_event, MemberEvent::Stripped(_)));
-        assert_eq!(self.get_room_infos().await.unwrap().len(), 0);
-        assert_eq!(self.get_stripped_room_infos().await.unwrap().len(), 1);
+        assert_eq!(self.get_room_infos().await.unwrap().len(), 1);
+        #[allow(deprecated)]
+        let stripped_rooms = self.get_stripped_room_infos().await?;
+        assert_eq!(stripped_rooms.len(), 1);
 
         let members = self.get_user_ids(room_id, RoomMemberships::empty()).await.unwrap();
         assert_eq!(members, vec![user_id.to_owned()]);
@@ -799,8 +955,10 @@ impl StateStoreIntegrationTests for DynStateStore {
 
         self.remove_room(room_id).await?;
 
-        assert!(self.get_room_infos().await?.is_empty(), "room is still there");
-        assert_eq!(self.get_stripped_room_infos().await?.len(), 1);
+        assert_eq!(self.get_room_infos().await?.len(), 1, "room is still there");
+        #[allow(deprecated)]
+        let stripped_rooms = self.get_stripped_room_infos().await?;
+        assert_eq!(stripped_rooms.len(), 1);
 
         assert!(self.get_state_event(room_id, StateEventType::RoomName, "").await?.is_none());
         assert!(
@@ -853,8 +1011,116 @@ impl StateStoreIntegrationTests for DynStateStore {
         self.remove_room(stripped_room_id).await?;
 
         assert!(self.get_room_infos().await?.is_empty(), "still room info found");
-        assert!(self.get_stripped_room_infos().await?.is_empty(), "still stripped room info found");
+        #[allow(deprecated)]
+        let stripped_rooms = self.get_stripped_room_infos().await?;
+        assert!(stripped_rooms.is_empty(), "still stripped room info found");
         Ok(())
+    }
+
+    async fn test_presence_saving(&self) {
+        let user_id = user_id();
+        let second_user_id = user_id!("@second:localhost");
+        let third_user_id = user_id!("@third:localhost");
+        let unknown_user_id = user_id!("@unknown:localhost");
+
+        // No event in store.
+        let mut user_ids = vec![user_id.to_owned()];
+        let presence_event = self.get_presence_event(user_id).await;
+        assert!(presence_event.unwrap().is_none());
+        let presence_events = self.get_presence_events(&user_ids).await;
+        assert!(presence_events.unwrap().is_empty());
+
+        // One event in store.
+        let mut changes = StateChanges::default();
+        changes.presence.insert(user_id.to_owned(), custom_presence_event(user_id));
+        self.save_changes(&changes).await.unwrap();
+
+        let presence_event = self.get_presence_event(user_id).await;
+        assert!(presence_event.unwrap().is_some());
+        let presence_events = self.get_presence_events(&user_ids).await;
+        assert_eq!(presence_events.unwrap().len(), 1);
+
+        // Several events in store.
+        let mut changes = StateChanges::default();
+        changes.presence.insert(second_user_id.to_owned(), custom_presence_event(second_user_id));
+        changes.presence.insert(third_user_id.to_owned(), custom_presence_event(third_user_id));
+        self.save_changes(&changes).await.unwrap();
+
+        user_ids.extend([second_user_id.to_owned(), third_user_id.to_owned()]);
+        let presence_event = self.get_presence_event(second_user_id).await;
+        assert!(presence_event.unwrap().is_some());
+        let presence_event = self.get_presence_event(third_user_id).await;
+        assert!(presence_event.unwrap().is_some());
+        let presence_events = self.get_presence_events(&user_ids).await;
+        assert_eq!(presence_events.unwrap().len(), 3);
+
+        // Several events in store with one unknown.
+        user_ids.push(unknown_user_id.to_owned());
+        let member_events = self.get_presence_events(&user_ids).await;
+        assert_eq!(member_events.unwrap().len(), 3);
+
+        // Empty user IDs list.
+        let presence_events = self.get_presence_events(&[]).await;
+        assert!(presence_events.unwrap().is_empty());
+    }
+
+    async fn test_display_names_saving(&self) {
+        let room_id = room_id!("!test_display_names_saving:localhost");
+        let user_id = user_id();
+        let user_display_name = "User";
+        let second_user_id = user_id!("@second:localhost");
+        let third_user_id = user_id!("@third:localhost");
+        let other_display_name = "Raoul";
+        let unknown_display_name = "Unknown";
+
+        // No event in store.
+        let mut display_names = vec![user_display_name.to_owned()];
+        let users = self.get_users_with_display_name(room_id, user_display_name).await.unwrap();
+        assert!(users.is_empty());
+        let names = self.get_users_with_display_names(room_id, &display_names).await.unwrap();
+        assert!(names.is_empty());
+
+        // One event in store.
+        let mut changes = StateChanges::default();
+        changes
+            .ambiguity_maps
+            .entry(room_id.to_owned())
+            .or_default()
+            .insert(user_display_name.to_owned(), [user_id.to_owned()].into());
+        self.save_changes(&changes).await.unwrap();
+
+        let users = self.get_users_with_display_name(room_id, user_display_name).await.unwrap();
+        assert_eq!(users.len(), 1);
+        let names = self.get_users_with_display_names(room_id, &display_names).await.unwrap();
+        assert_eq!(names.len(), 1);
+        assert_eq!(names.get(&user_display_name).unwrap().len(), 1);
+
+        // Several events in store.
+        let mut changes = StateChanges::default();
+        changes.ambiguity_maps.entry(room_id.to_owned()).or_default().insert(
+            other_display_name.to_owned(),
+            [second_user_id.to_owned(), third_user_id.to_owned()].into(),
+        );
+        self.save_changes(&changes).await.unwrap();
+
+        display_names.push(other_display_name.to_owned());
+        let users = self.get_users_with_display_name(room_id, user_display_name).await.unwrap();
+        assert_eq!(users.len(), 1);
+        let users = self.get_users_with_display_name(room_id, other_display_name).await.unwrap();
+        assert_eq!(users.len(), 2);
+        let names = self.get_users_with_display_names(room_id, &display_names).await.unwrap();
+        assert_eq!(names.len(), 2);
+        assert_eq!(names.get(&user_display_name).unwrap().len(), 1);
+        assert_eq!(names.get(&other_display_name).unwrap().len(), 2);
+
+        // Several events in store with one unknown.
+        display_names.push(unknown_display_name.to_owned());
+        let names = self.get_users_with_display_names(room_id, &display_names).await.unwrap();
+        assert_eq!(names.len(), 2);
+
+        // Empty user IDs list.
+        let names = self.get_users_with_display_names(room_id, &[]).await;
+        assert!(names.unwrap().is_empty());
     }
 }
 
@@ -980,6 +1246,18 @@ macro_rules! statestore_integration_tests {
             let store = get_store().await?.into_state_store();
             store.test_room_removal().await
         }
+
+        #[async_test]
+        async fn test_presence_saving() {
+            let store = get_store().await.expect("creating store failed").into_state_store();
+            store.test_presence_saving().await;
+        }
+
+        #[async_test]
+        async fn test_display_names_saving() {
+            let store = get_store().await.expect("creating store failed").into_state_store();
+            store.test_display_names_saving().await;
+        }
     };
 }
 
@@ -1034,10 +1312,10 @@ fn custom_stripped_membership_event(user_id: &UserId) -> Raw<StrippedRoomMemberE
 }
 
 fn membership_event() -> Raw<SyncRoomMemberEvent> {
-    custom_membership_event(user_id(), event_id!("$h29iv0s8:example.com").to_owned())
+    custom_membership_event(user_id(), event_id!("$h29iv0s8:example.com"))
 }
 
-fn custom_membership_event(user_id: &UserId, event_id: OwnedEventId) -> Raw<SyncRoomMemberEvent> {
+fn custom_membership_event(user_id: &UserId, event_id: &EventId) -> Raw<SyncRoomMemberEvent> {
     let ev_json = json!({
         "type": "m.room.member",
         "content": RoomMemberEventContent::new(MembershipState::Join),
@@ -1045,6 +1323,17 @@ fn custom_membership_event(user_id: &UserId, event_id: OwnedEventId) -> Raw<Sync
         "origin_server_ts": 198,
         "sender": user_id,
         "state_key": user_id,
+    });
+
+    Raw::new(&ev_json).unwrap().cast()
+}
+
+fn custom_presence_event(user_id: &UserId) -> Raw<PresenceEvent> {
+    let ev_json = json!({
+        "content": {
+            "presence": "online"
+        },
+        "sender": user_id,
     });
 
     Raw::new(&ev_json).unwrap().cast()