+++ crates/matrix-sdk-sqlite/migrations/state_store/002_a_create_new_room_info.sql
@@ -1,0 +1,6 @@
+-- The rooms are not filtered by stripped state anymore but by state.
+CREATE TABLE "new_room_info" (
+    "room_id" BLOB PRIMARY KEY NOT NULL,
+    "state" BLOB NOT NULL,
+    "data" BLOB NOT NULL
+);