+++ crates/matrix-sdk-crypto/CHANGELOG.md
@@ -1,5 +1,7 @@
 # v0.7.0
 
+- Add support for the `hkdf-hmac-sha256.v2` SAS message authentication code.
+
 - Ensure that the correct short authentication strings are used when accepting a
   SAS verification with the `Sas::accept()` method.